refactor: class ExtractorPhind
