+++ src/tab/exportProcessor/scraper/extractor/ExtractorPhindAgent.js
@@ -1,8 +1,8 @@
 import {setFileHeader} from "../../formatter/formatMarkdown";
 import {sleep} from "../../../../common/utils";
-import {Extractor} from "./Extractor";
+import ExtractorPhind from "./ExtractorPhind";
 
-export default class ExtractorPhindSearch extends Extractor {
+export default class ExtractorPhindSearch extends ExtractorPhind {
   constructor(domain) {
     super(domain);
   }
@@ -66,16 +66,6 @@ export default class ExtractorPhindSearch extends Extractor {
     }
 
     return markdown;
-  }
-
-  getPageTitle() {
-    const searchFirstMsg = document.querySelector('[name^="answer-"] > div > div > span');
-    const agentFirstMsg = document.querySelector('[tabindex="0"]');
-    return searchFirstMsg !== null && searchFirstMsg.innerHTML !== ""
-      ? searchFirstMsg.innerHTML
-      : agentFirstMsg
-        ? agentFirstMsg.innerText.replace(/\u00A0/g, " ")
-        : "";
   }
 
   getPageSource() {
